# internal/swagger/docs/swagger.yaml
definitions:
  handlers.InputCourse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  handlers.InputPerson:
    properties:
      age:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      type:
        type: string
    type: object
  handlers.ResponseErr:
    properties:
      error:
        type: string
    type: object
  handlers.ResponseMsg:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      description: A test endpoint to get everything up and running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Hello, World!
      tags:
      - test
  /courses:
    get:
      consumes:
      - application/json
      description: List all courses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: List all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a course
      parameters:
      - description: Course Object
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/handlers.InputCourse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Create a course
      tags:
      - courses
  /courses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a course by their ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Delete course by ID
      tags:
      - courses
    get:
      consumes:
      - application/json
      description: Get a course by their ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Get course by ID
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Update a course by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Course Object
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/handlers.InputCourse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Update a course by id
      tags:
      - courses
  /health-check:
    get:
      consumes:
      - application/json
      description: Health check response
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
      summary: Health check response
      tags:
      - health-check
  /people:
    get:
      consumes:
      - application/json
      description: List all people
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: List all people
      tags:
      - people
    post:
      consumes:
      - application/json
      description: Create a person
      parameters:
      - description: Person Object
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/handlers.InputPerson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Create a person
      tags:
      - people
  /people/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a person by their ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Delete person by ID
      tags:
      - people
    get:
      consumes:
      - application/json
      description: Get a person by their ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Get person by ID
      tags:
      - people
    put:
      consumes:
      - application/json
      description: Update a person by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Person Object
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/handlers.InputPerson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ResponseMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ResponseErr'
      summary: Update a person by id
      tags:
      - people
swagger: "2.0"